batadaze/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.20
python-dotenv==1.0.1 